requirements.txt
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
h3==3.7.7
idna==3.7
numpy==2.0.1
pillow==10.4.0
pycparser==2.22
pytz==2024.1
requests==2.32.3
timezonefinder==6.5.2
urllib3==2.2.2
